RPT-8: Extract common part from PageForDate and hide the particular implementation behind the interface.
