+++ Src/Clock/Clock.hpp
@@ -12,7 +12,6 @@
 
 #pragma once
 
-#include "pico/stdlib.h"
 #include "FreeRTOS.h"
 #include "queue.h"
 #include "task.h"
