+++ Src/UserInterface/Pages/IPage.hpp
@@ -1,0 +1,15 @@
+#pragma once
+
+enum class EventProcessingResult {
+    Continue,
+    Cancel,
+    Apply,
+};
+
+class IPage
+{
+public:
+    virtual ~IPage() = default;
+    virtual void Render() = 0;
+    virtual EventProcessingResult ProcessMenuEvent(MenuEvent event) = 0;
+};
